pyproject.toml
[project]
name = "ml-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "joblib>=1.5.2",
    "pre-commit>=4.3.0",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.1",
    "uvicorn>=0.35.0",
]
